Fixed Bug Caused When Pistol was in gun inventory
The if statement added to fix the previous bug created a new bug if the
player had a pistol in the gun inventory. I moved the new if statement
inside the pistol check if statement to correct this.
WARNING - Code is not producing a second chest, I am still trying to
investigate why.